[김민지] 폴더 생성 후 이동
