--- README.md
@@ -1,2 +1,0 @@
-# 2021-02-Spring-Study
-[2021 GDSC Sookmyung 2기] 1분기 스터디(Spring)를 위한 저장소
