--- freelec-springboot2-webservice/src/test/java/com/mjkim/springboot/web/HelloControllerTest.java
@@ -1,46 +1,0 @@
-package com.mjkim.springboot.web;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.context.junit4.SpringRunner;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.ResultActions;
-
-import static org.hamcrest.Matchers.is;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-
-@RunWith(SpringRunner.class) // JUnit에 내장된 실행자가 아니라 SpringRunner라는 스프링 실행자 사용
-@WebMvcTest(controllers = HelloController.class) // Web(Spring MVC)에 집중할 수 있는 어노테이션
-public class HelloControllerTest {
-
-    @Autowired // 스프링이 관리하는 Bean을 주입받음
-    private MockMvc mvc; // 웹 API를 테스트할 때 사용, 스프링 MVC 테스트의 시작점
-
-    @Test
-    public void hello가_리턴된다() throws Exception {
-        String hello = "hello";
-
-        mvc.perform(get("/hello")) // MockMvc를 통해 /hello 주소로 HTTP GET 요청
-                .andExpect(status().isOk()) // HTTP Header의 Status 검증
-                .andExpect(content().string(hello)); // 응답 본문 내용 검증, Controller에서 "hello"를 리턴하기 때문에 이 값이 맞는지 검증
-    }
-
-    @Test
-    public void helloDto가_리턴된다() throws Exception {
-        String name = "hello";
-        int amount = 1000;
-
-        mvc.perform(get("/hello/dto")
-                .param("name", name) // API 테스트 시 사용될 요청 파라키터 설정, 값은 String만 허용(문자열로 변경 필요)
-                .param("amount", String.valueOf(amount))
-        )
-                .andExpect(status().isOk())
-                .andExpect(jsonPath("$.name", is(name))) // jsonPath: JSOM 응답값을 필드 별로 검증
-                .andExpect(jsonPath("$.amount", is(amount)));
-    }
-}
